+++ changelog.md
@@ -75,6 +75,7 @@ Next, build a demo app on top of the app infrastructure and database that is in 
 * Add a page to list available surveys.
 * Add a utility to create ids from strings. We'll use titles and questions to create ids.
 * Add a `survey` endpoint that understands survey ids and which survey to pick up.
+* Add an intro page for begining a session
 
 I used a couple blog posts about how to set up Express and postgres:
 * https://medium.freecodecamp.org/how-to-make-create-react-app-work-with-a-node-backend-api-7c5c48acb1b0